refactor(mobile): partir perfil_tab.dart en 4 modulos (fase 5a-2)
- perfil_tab.dart       328  pantalla + state (carga perfil, foto, build)
- cabecera_perfil.dart  174  header con foto/iniciales + badge activo
- secciones_perfil.dart 156  titulo seccion + tarjeta + fila + toggle tema
- dialogs_perfil.dart   286  dialog editar info medica + cambiar password

Archivo original (927 lineas) movido a subcarpeta perfil/.
Sin cambios funcionales. flutter analyze: 0 issues.
