+++ mobile/lib/screens/student/student_home_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:yoltec_mobile/screens/student/citas_tab.dart';
 import 'package:yoltec_mobile/screens/student/inicio_tab.dart';
-import 'package:yoltec_mobile/screens/student/perfil_tab.dart';
+import 'package:yoltec_mobile/screens/student/perfil/perfil_tab.dart';
 import 'package:yoltec_mobile/screens/student/recetas_tab.dart';
 import 'package:yoltec_mobile/screens/student/widgets/student_widgets.dart';
 import 'package:yoltec_mobile/services/auth_service.dart';
